ABIOS-692 Changed Optimizely token and projectID, removed weird UIKit

+++ Ambatana/src/ViewControllers/Navigation/TabBarController.swift
@@ -189,7 +189,7 @@ public final class TabBarController: UITabBarController, NewSellProductViewContr
                     if shouldSelectVC {
                         
                         // Change the tab
-                        selectedIndex = UInt(vcIdx)
+                        selectedIndex = vcIdx
                         
                         // Pop the navigation back to root
                         navBarCtl.popToRootViewControllerAnimated(false)
